refactor timestampables of entities into traid - move properties created_time, last_updated_time, etc. to its own trait - remove attributes and methods from entity classes (which are applied by traid) - apply traid to according entity classes (database, page and block) - add created_by, last_updated_by to this new trait
