+++ src/Entities/Blocks/Block.php
@@ -5,6 +5,7 @@ namespace FiveamCode\LaravelNotionApi\Entities\Blocks;
 use DateTime;
 use FiveamCode\LaravelNotionApi\Entities\Entity;
 use FiveamCode\LaravelNotionApi\Exceptions\HandlingException;
+use FiveamCode\LaravelNotionApi\Traits\TimestampableEntity;
 use Illuminate\Support\Arr;
 
 /**
@@ -12,6 +13,8 @@ use Illuminate\Support\Arr;
  */
 class Block extends Entity
 {
+    use TimestampableEntity;
+
     /**
      * @var string
      */
@@ -38,16 +41,6 @@ class Block extends Entity
     protected string $text = '[warning: unsupported in notion api]';
 
     /**
-     * @var DateTime
-     */
-    protected DateTime $createdTime;
-
-    /**
-     * @var DateTime
-     */
-    protected DateTime $lastEditedTime;
-
-    /**
      * @param  array  $responseData
      *
      * @throws HandlingException
@@ -69,8 +62,7 @@ class Block extends Entity
         $this->fillType();
         $this->fillRawContent();
         $this->fillHasChildren();
-        $this->fillCreatedTime();
-        $this->fillLastEditedTime();
+        $this->fillTimestampableProperties();
     }
 
     private function fillType(): void
@@ -124,22 +116,6 @@ class Block extends Entity
     public function hasChildren(): bool
     {
         return $this->hasChildren;
-    }
-
-    /**
-     * @return DateTime
-     */
-    public function getCreatedTime(): DateTime
-    {
-        return $this->createdTime;
-    }
-
-    /**
-     * @return DateTime
-     */
-    public function getLastEditedTime(): DateTime
-    {
-        return $this->lastEditedTime;
     }
 
     public function getContent()
